refactor: split HeaderV2 into focused sibling files under HeaderV2/
The single 536-line `HeaderV2.tsx` mixed: a constants block, the info
sidebar builder, custom-filters dialog state + modal rendering + nav
side-effects, the task-filters sidebar builder (with hover-revealed
edit/delete actions per custom filter), the top-level sidebar layout,
the camunda-tools options object, the useC3NavigationV2 call, the
LanguageSelector component, and the ThemeSelector component.

Split into one folder by responsibility:

  HeaderV2/
  ├── index.tsx                     ← composition + render (~120 lines)
  ├── LanguageSelector.tsx
  ├── ThemeSelector.tsx
  ├── info-sidebar-items.ts         ← Documentation / Academy / Feedback / Forum
  ├── use-custom-filters-dialog.tsx ← localStorage read + modal lifecycle + nav side-effects
  ├── use-task-filters-sidebar.tsx  ← built-in filters + per-custom-filter row with hover actions
  ├── use-sidebar-children.tsx      ← top-level Tasks group + Processes item
  └── use-camunda-tools-config.tsx  ← notifications / info / user panels

The composition layer in `index.tsx` is now linear: hooks return data
or rendered modals, the C3 call assembles them. No descriptor-building
in the component body. No behaviour change; 38/38 Header tests stay
green.
